Ability to create a map of hispanic percentage based on county code. Need to automate state collection

+++ auto_run.py
@@ -1,10 +1,11 @@
 import census_df as cdf
 import folder_paths as fpaths
+from os import path
 
 
 # To dataframe
-def to_json(path=fpaths.alabama_path):
-    cdf.read(path)
+def to_json(spath=fpaths.alabama_path):
+    cdf.read(spath)
 
 
 # Press the green button in the gutter to run the script.
@@ -12,10 +13,13 @@ if __name__ == '__main__':
     for state_path in fpaths.location_paths:
         state_name = state_path.split("/")[1]
 
-        print("Processing", state_name)
+        if not path.exists("State_JSONs/" + state_name + ".json"):
+            print("Processing", state_name)
 
-        print("\tStarting to create dataframe")
+            print("\tStarting to create dataframe")
 
-        to_json(state_path)
+            to_json(state_path)
 
-        print(state_name, "Processed")
+            print(state_name, "Processed")
+        else:
+            print(state_name, "Already exists, skipping.")
